+++ controllers/SitehcnController.php
@@ -115,6 +115,11 @@ class SitehcnController extends Controller
         }
     }
 
+    //habilitar ajax
+    public function beforeAction($action) {
+        $this->enableCsrfValidation = false;
+        return parent::beforeAction($action);
+    }
     /**
      * Deletes an existing Site model.
      * If deletion is successful, the browser will be redirected to the 'index' page.
@@ -123,9 +128,17 @@ class SitehcnController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        
+        $projeto = Yii::$app->db->createCommand('SELECT id FROM projeto WHERE site='.$id)->queryScalar();
+        if(empty($projeto)){
+            $this->findModel($id)->delete();
+        }
+        else{
+            //botar uma mensagem (setFlash)
+            echo 'Existe um projeto utilizando este site';
+        }
 
-        return $this->redirect(['index']);
+        return $this->redirect(['create']);
     }
 
     /**
